agents/cloudrun/requirements.txt
google-adk
python-dotenv
fastapi
uvicorn
aiosqlite

